refactor: extract setup flow mixin from config flow
